Delete job always taking "scaleTimeout" time to delete a job
countDownLatch.countDown() is never getting called.
job.getStatus().getActive() is returning null if job is scalled down to 0
The reason here is that the api server does not send "active" field inside the response JSON.
Hence jackson deserializes it to null. Only if the job is active is the active field set to some int.
Hence delete job always takes getConfig().getScaleTimeout() amount of time to terminate.