Fix project creation for Openshift 3.6+
This fixes the following failure:

Message: the API version in the data (v1) does not match the expected
API version (project.openshift.io/v1).

Since we are now using API Groups (when they are available), it seems
the project creation was not updated to use the new API group name,
leading to the error above, due to mixing usage of old 'v1' API and
new 'project.openshift.io/v1' API group.

So, this patch updates the project creation to use the new API group
before actually making the API call to Openshift.

Closes #829.