Add delay in subsequent tests to handle resource depletion
Right now in regression tests, we fire up too many pods on this very small
CircleCI cluster which results in cluster being not able to process pods; So pods
not really come in Ready state. It keeps throwing `Insufficient pods` error.

So add some delay in tests for graceful deletion of resources.