Prevent IOException if certificate file has trailing empty lines
As described in issue #987 it is possible for certificate files that include trailing empty lines to trigger IOExceptions because the CertUtils createTrustStore() method continues to work with the pemInputStream even though all of its certificates have been processed. It looks like this scenario is not being caught in the CertUtilsTest tests because they do not actually use the CertUtils.getInputStreamFromDataOrFile() method to get an input stream from the test PEM files but instead used getClass().getResourceAsStream().