Add support for API groups to attribute extractor.
Custom resource definitions and many APIs live in API groups other than
"extensions". The current regex doesn't accept these groups and leads to
silent failures in the mock behavior because the path can't be parsed.

Technically, these API groups should be part of the attributes, but they
can't be extracted from a HasMetadata interface and I suspect we'd be
introducing matching failures if we added them for URL-generated
attributes but not object-generated attributes.