Remove the test because its failing on CI.
If we keep that test, it will return different URl for
every other CI. If we do some changes in test, it will
be similar to other tests like setting masterUrl and all.
This started happening after PR https://github.com/fabric8io/kubernetes-client/pull/1086
Asserting not null and is URL is the only solution
but does not make any difference as we have another tests for this func.