Add support for label selectors in the mock server (#1158)
* Support label selectors in mock server

This supports simple label selectors (eg foo=bar). The != operator would
go beyond the limits of the AttributeSet-based matching in
CrudDispatcher.

There was a test using labels in ConfigMapCrudTest, but it was only
passing because, at the point of execution, all resources match the
label selector. This change moves changes the order of assertions so
that the effect of the label selector is observable.

This should fix #1125, although I haven't tested that code directly.

* Fix PodCrudTest's use of label selector

It was only passing before because the label selector was ignored.

* Improve flexibility of URL parsing

As recommended by @lordofthejars. I renamed PROTOCOL to SCHEME to match
the convention of HttpUrl.