Fix tracking of watched resource versions (#1215)
* Fix tracking of watched resource versions

Fixes #1214

resourceVersions are Strings, so comparing them with compareTo leads to
unexpected results ("9".compareTo("10") is > 0.) Also since the versions
are supposed to be opaque we should not be trying to read any meaning
or ordering into them.

Since messages are processed in order we should just always store the
last processed resourceVersion.

I'm not sure this even needs to use an Atomic* type, but it seems safer
to just leave that in place.

* Remove test job running on CircleCI, which was added due to failing Jenkins