fix #1495: avoid runtime dependency on Commons Collections
The fix for issue #1468 introduced a runtime dependency
on Commons Collections just for `MapUtils.EMPTY_MAP`.
However, depending on Fabric8 Kubernetes Client doesn't
transitively bring in Commons Collections, so invoking
the respective codepath results in `NoClassDefFoundError`.

This commit replaces the usage of `MapUtils.EMPTY_MAP`
by `Collections.emptyMap()`, which does the same and
is available in the Java Collections Framework.