chore: cleanup log4j dependency from project
The SLF4J is used as logging dependency, log4j is used only on tests, so this replaces the log4j declaration for slf4j-simple in the test scope.

This cleanup helps to reduce the dependencies used in the project and keep a single implementation in place.

This also removes the unnecessary dependency on jul-to-slf4j, which ends up being transitively included in projects that use kubernetes-client.