feat: let CRDGenerator clients handle output
This is meant so that clients of the CRD generation API can decide what
to output and how. Information that wasn't previously exposed is now
made available on CRDInfo for interested clients.

Note also that `CRDInfo.getVersion` has been deprecated in favor of
`getCrdSpecVersion` to clarify which version we're talking about: the
CRD spec, not the CRD version.

Fixes #3366