fix: sundrio path traversal issue
When compiled in Java 8, there's an error caused by sundrio
not being able to traverse the full dependency graph of
OkHttpClient.

Most likely OkHttp was compiled with JDK11+ and has
references that are not available for local compilation.

Sundrio faces this issue since when traversing the OkHttp AST
for the return type it ends up analyzing types that are not
available in the current JDK.