Fix deserialization of interfaces in the generated model
As part of https://github.com/fabric8io/kubernetes-client/pull/3594, the go generator will generate the interfaces and their implementations. Example:

Go type `Tracing_CustomTag` that uses a field `Type` which is an interface `isTracing_CustomTag_Type`:
```go
type Tracing_CustomTag struct {
	Type                 isTracing_CustomTag_Type `protobuf_oneof:"type"`
}
```

Where the interface is defined in:

```go
type isTracing_CustomTag_Type interface {
	isTracing_CustomTag_Type()
	MarshalTo([]byte) (int, error)
	Size() int
}
```

The implementations of this interface are defined in:

```go
type Tracing_CustomTag_Literal struct {
	Literal *Tracing_Literal `protobuf:"bytes,1,opt,name=literal,proto3,oneof" json:"literal,omitempty"`
}
type Tracing_CustomTag_Environment struct {
	Environment *Tracing_Environment `protobuf:"bytes,2,opt,name=environment,proto3,oneof" json:"environment,omitempty"`
}
type Tracing_CustomTag_Header struct {
	Header *Tracing_RequestHeader `protobuf:"bytes,3,opt,name=header,proto3,oneof" json:"header,omitempty"`
}
```

The go generator will handle the generation of the Java interface and their implementations. However, this PR fixes the deserialization of the fields that are interfaces. This is when using the client to load a model from the cluster and this model has an interface. 

At the moment, the field that is an interface will be deserialized to null, so we need to annotate the interface with:

```java
@JsonInclude(JsonInclude.Include.NON_NULL)
@JsonTypeResolver(io.fabric8.kubernetes.model.jackson.UnwrappedTypeResolverBuilder.class) (1)
@JsonTypeInfo(use = JsonTypeInfo.Id.DEDUCTION) (2)
@JsonSubTypes({
    @JsonSubTypes.Type(TracingCustomTagEnvironment.class),
    @JsonSubTypes.Type(TracingCustomTagHeaderName.class),
    // ...
})
public interface IsTracing_CustomTag_Type {
}
```

Using the DEDUCTION mode (2), the jackson deserializer will deduct what is the right implementation to pick.
The problem is that this annotation also affects the serialization, so we need the annotation in (1) to ignore these annotations when serialization.