Fix #3568: Pod upload file fails if file is directly in root path
Pod file upload seems be be creating invalid exec command for upload if
the file is directly in root path e.g. `/cp.log`. We create parent
directory of file by creating a substring based on last index of `/`.
This returns an empty string for a string like `/cp.log`.

Assume directory path to be `/` when directory is returned as empty string.

Signed-off-by: Rohan Kumar <rohaan@redhat.com>