Apply minor optimization for to decrease allocations
When performing a small profiling of a sample Operator SDK application,
I noticed that BeanPropertyWriterDelegate was leading to a lot of
allocations due to the use of lambdas.
Lambdas are great, but can lead to performance issues when used on the
hot path of the code, which this class seems to be.
Furthermore, the change does not reduce the readability of the code

(cherry picked from commit 9e057d2f96119f1ab3e99d952382c8f7cacd7a17)