Fixes #3880: Synchronize access to map in KubernetesCrudDispatcher
The CrudDispatcher map is a synchronized map. The implementation of the standard map operations are already internally synchronized, but it explicitly states that iterating the map in any way should be synchronized. Occasionally in a multi-threaded scenario access to the map results in a ConcurrentModificationException because of a lack of synchronization.

(cherry picked from commit 22874f0f032da7e7f8d31f0e232d91ff3448e0c0)
Signed-off-by: Marc Nuri <marc@marcnuri.com>