Fix use case with FIPS library
The code as it was does not work with FIPS, because the class name is different. The check for BCFIPS shows the coder expected this use case, but having the instantiation of the standard BC provider class hard coded inside the Callable<> made the NoClassDefFoundError unavoidable in the FIPS case.