fix (kubernetes-client) : Change upload tarball location to target upload directory
Using `/tmp` as upload directory can be problematic for scenarios where
`/tmp` is read only. We can directly upload the intermediate upload
tarball to the directory where we are uploading the content.

In case of uploading a file to a Pod, temporary tarball would be uploaded in
the parent directory of that file.

In case of uploading a directory to a Pod, temporary tarball would be
uploaded in the same directory that's specified for upload path.

Signed-off-by: Rohan Kumar <rohaan@redhat.com>