CRD generator implementation that supports Jakarta and Swagger
This change adds an alternative CRD generator implementation that is
based on
[victools/jsonschema-generator](https://github.com/victools/jsonschema-generator).

The victools generator allows modules to be registered to customize
schema generation, and there are modules that provide support for
Jakarta Validation and Swagger annotations.

These changes are being contributed into test scope as a prototype,
since no effort has been made to integrate it with the existing CRD
generator API. It is a totally separate implementation that is not at
parity with the existing generator. The intent is that someone with more
knowledge of the Fabric8 codebase can incorporate some aspects of it
into the real product.