feat(openapi): go openapi schema + json openapi schema generation
The generator is capable of parsing go types to generate
a zz_generated.openapi.go file containing the OpenAPI schema definitions.
The file is generated using kube-openapi openapi-gen programmatically
as a go-generator script.

The file is further processed to generate the openshift-generated.json file
that contains the final OpenAPI schema definitions augmented with x-fabric8-info
that includes the additional information to be able to infer group, version, kind,
and so on.

Signed-off-by: Marc Nuri <marc@marcnuri.com>